% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Fetch a list of areas}
\usage{
area(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Optional parameter to enable verbose logging (default = \code{FALSE}).}
}
\value{
The areas.
}
\description{
Fetch a list of areas
}
\examples{
areas <- area(verbose = FALSE)
}
