% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Create a checklist.}
\usage{
checklist(scientificname = NULL, taxonid = NULL, datasetid = NULL,
  nodeid = NULL, startdate = NULL, enddate = NULL,
  startdepth = NULL, enddepth = NULL, geometry = NULL,
  verbose = FALSE)
}
\description{
Create a checklist.
}
