% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{dna_records}
\alias{dna_records}
\title{Extract DNA records from occurrence data with a dna column.}
\usage{
dna_records(df, fields = "id")
}
\arguments{
\item{df}{the occurrence dataframe.}

\item{fields}{columns from the occurrence dataframe to include.}
}
\value{
The DNA records.
}
\description{
Extract DNA records from occurrence data with a dna column.
}
