% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_citation}
\alias{generate_citation}
\title{Generate a citation from metadata elements.}
\usage{
generate_citation(title, published, url, contacts)
}
\arguments{
\item{title}{the dataset title.}

\item{published}{the dataset published date.}

\item{url}{the dataset url.}

\item{contacts}{the dataset contacts as a dataframe.}
}
\value{
A citation string.
}
\description{
Generate a citation from metadata elements.
}
