% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Distribution}
\usage{
get_distribution(identifier, table = "dist_sp_5deg", maxFeatures = NULL,
  ...)
}
\arguments{
\item{identifier}{species name}

\item{table}{table}

\item{maxFeatures}{maximum number of features}

\item{...}{additional arguments for \code{\link{wfs_request}}}
}
\value{
distribution data
}
\description{
Retrieves distribution data
}

