% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_geometry}
\alias{get_geometry}
\title{Get a WKT geometry by drawing on a map.}
\usage{
get_geometry(provider_tiles = "Esri.WorldGrayCanvas")
}
\arguments{
\item{provider_tiles}{the base map provider.}
}
\description{
Get a WKT geometry by drawing on a map.
}
