% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occurrences.R
\name{get_occurrences}
\alias{get_occurrences}
\title{Occurrences}
\usage{
get_occurrences(identifiers, id = "aphia", filter = NULL, where = NULL,
  ...)
}
\arguments{
\item{identifiers}{vector of identifiers, these can be species names, AphiaIDs, OBIS Taxon IDs, or storedpaths}

\item{id}{which id to use when the identifier is an id, one of \code{aphia}, \code{obis}}

\item{filter}{list of filters}

\item{where}{explicit where clause}

\item{...}{additional arguments for \code{\link{wfs_request}}}
}
\value{
occurrence data
}
\description{
Retrieves occurrence data
}

