% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Fetch taxonomic groups}
\usage{
group(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Optional parameter to enable verbose logging (default
= \code{FALSE}).}
}
\value{
The taxonomic groups.
}
\description{
Fetch taxonomic groups
}
\seealso{
\code{\link{occurrence}} \code{\link{checklist}} \code{\link{node}}
  \code{\link{area}} \code{\link{dataset}} \code{\link{taxon}}
}
