% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{leafletmap}
\alias{leafletmap}
\title{Create a Leaflet map.}
\usage{
leafletmap(data, popup = NULL)
}
\arguments{
\item{data}{occurrence records you want to map, minimal required fields are
\code{decimalLongitude} and \code{decimalLatitude}.}

\item{popup}{Column to be used for popups.}
}
\value{
HTML widget object.
}
\description{
Create a Leaflet map.
}
\seealso{
\code{\link{occurrence}}
}
