% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_ggplot}
\alias{map_ggplot}
\title{Create a ggplot2 map.}
\usage{
map_ggplot(data, color = "#ff3399")
}
\arguments{
\item{data}{the occurrences from \code{occurrence()}.}

\item{color}{color to be used for the dots.}
}
\description{
Create a ggplot2 map.
}
