% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Create a leaflet map.}
\usage{
map_leaflet(data, color = "#ff3399",
  provider_tiles = "OpenStreetMap.BlackAndWhite", popup = function(x) {
      x["id"] })
}
\description{
Create a leaflet map.
}
