% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_occurrences.R
\name{map_occurrences}
\alias{map_occurrences}
\title{Occurrence map}
\usage{
map_occurrences(data, color = "tname")
}
\arguments{
\item{data}{data frame from \code{\link{get_occurrences}}}

\item{color}{attribute to be used for color coding}
}
\description{
Creates a map from occurrence data
}

