% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_summaries.R
\name{map_summaries}
\alias{map_summaries}
\title{Summary map}
\usage{
map_summaries(data, summary = "shannon", style = NULL)
}
\arguments{
\item{data}{data frame from \code{\link{get_summaries}} or \code{\link{get_hexsummaries}}}

\item{summary}{summary name to be used, either \code{simpson}, \code{shannon}, \code{es}, \code{s}, \code{chao2_unbiased_biota}, or \code{completeness_biota}}

\item{style}{list with breaks, labels, and colors vectors}
}
\description{
Creates a map from summary data
}

