% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mof.R
\name{measurements}
\alias{measurements}
\title{Extract measurements or facts from occurrence data with a mof column.}
\usage{
measurements(df, fields = c("id"))
}
\arguments{
\item{df}{the occurrence dataframe.}

\item{fields}{columns from the occurrence dataframe to include.}
}
\value{
The measurements.
}
\description{
Extract measurements or facts from occurrence data with a mof column.
}
