% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{Fetch nodes}
\usage{
node(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Optional parameter to enable verbose logging (default
= \code{FALSE}).}
}
\value{
The nodes.
}
\description{
Fetch nodes
}
\seealso{
\code{\link{occurrence}} \code{\link{checklist}} \code{\link{group}}
  \code{\link{area}} \code{\link{dataset}} \code{\link{taxon}}
}
