% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa}
\alias{taxa}
\title{Generate taxon lists.}
\usage{
taxa(scientificname = NULL, year = NULL, obisid = NULL, aphiaid = NULL,
  resourceid = NULL, startdate = NULL, enddate = NULL, geometry = NULL,
  qc = NULL, verbose = FALSE)
}
\arguments{
\item{geometry}{A wkt geometry string.}

\item{qc}{A vector of qc numbers you want to filter out.}
}
\value{
The taxon list.
}
\description{
Generate taxon lists.
}

