% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon_common}
\alias{taxon_common}
\title{Fetch common names.}
\usage{
taxon_common(verbose = FALSE)
}
\arguments{
\item{obisid}{numeric. OBIS identifier for the taxon.}

\item{verbose}{logical. Optional parameter to enable verbose logging (default = \code{FALSE}).}
}
\value{
The areas.
}
\description{
Fetch common names.
}
\seealso{
\code{\link{taxon_common}} \code{\link{occurrence}} \code{\link{checklist}} \code{\link{node}} \code{\link{area}}
  \code{\link{dataset}} \code{\link{group}}
}
