% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unnest_extension}
\alias{unnest_extension}
\title{Extract extension records from occurrence data with nested extension column.}
\usage{
unnest_extension(df, extension, fields = "id")
}
\arguments{
\item{df}{the occurrence dataframe.}

\item{extension}{the extension type (e.g. \code{MeasurementOrFact}, \code{DNADerivedData}).}

\item{fields}{columns from the occurrence dataframe to include.}
}
\value{
The extension records.
}
\description{
Extract extension records from occurrence data with nested extension column.
}
