% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/request.R
\name{wfs_request}
\alias{wfs_request}
\title{WFS request}
\usage{
wfs_request(typeName, version = "1.0.0", request = "GetFeature",
  outputFormat = "csv", maxFeatures = 100, bbox = c(-180, -90, 180, 90),
  propertyName = NULL, viewparams = NULL)
}
\arguments{
\item{typeName}{feature type name}

\item{version}{service version}

\item{request}{operation name}

\item{outputFormat}{output format, only csv is supported}

\item{maxFeatures}{maximum number of features}

\item{bbox}{bounding box as \code{c(min_longitude, min_latitude, max_longitude, max_latitude)}}

\item{propertyName}{vector of attributes to be returned}

\item{viewparams}{viewparams}
}
\value{
data frame
}
\description{
Makes a WFS request to OBIS
}

