% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\title{Generate checklists.}
\usage{
checklist(scientificname = NULL, year = NULL, obisid = NULL,
  aphiaid = NULL, resourceid = NULL, areaid = NULL, startdate = NULL,
  enddate = NULL, startdepth = NULL, enddepth = NULL, geometry = NULL,
  qc = NULL, verbose = FALSE)
}
\arguments{
\item{scientificname}{The full scientific name, with authorship and date
information if known.}

\item{year}{The year in which the Event occurred.}

\item{obisid}{The OBIS identifier of the species.}

\item{aphiaid}{The WoRMS identifier of the species.}

\item{resourceid}{The dataset identifier.}

\item{areaid}{The OBIS area identifier (country, marine world heritage site,
ABNJ, EBSA, ...). See also \code{\link{area}} for the list areas.}

\item{startdate}{The earliest date on which the Event occurred.}

\item{enddate}{The latest date on which the Event occurred.}

\item{startdepth}{The minimum depth below the sea surface.}

\item{enddepth}{The maximum depth below the sea surface.}

\item{geometry}{A wkt geometry string.}

\item{qc}{A vector of quality control flags you want to filter on. List of
\link[=qc]{QC flags}.}

\item{verbose}{logical. Optional parameter to enable verbose logging (default
= \code{FALSE}).}
}
\value{
The checklist.
}
\description{
Generate checklists.
}
\seealso{
\code{\link{occurrence}} \code{\link{dataset}} \code{\link{area}}
}
