% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Fetch dataset metadata.}
\usage{
dataset(id, verbose = FALSE)
}
\arguments{
\item{id}{numeric. One or more dataset ids.}

\item{verbose}{logical. Optional parameter to enable verbose logging (default
= \code{FALSE}).}
}
\value{
The dataset metadata records.
}
\description{
Fetch dataset metadata.
}
