% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaries.R
\name{get_hexsummaries}
\alias{get_hexsummaries}
\title{Summaries (hexagonal grid)}
\usage{
get_hexsummaries(typeName = "hexgrid4", maxFeatures = NULL, ...)
}
\arguments{
\item{typeName}{feature type name, one of \code{hexgrid1}, \code{hexgrid2}, \code{hexgrid3}, \code{hexgrid4}, \code{hexgrid5}, \code{hexgrid6}}

\item{maxFeatures}{maximum number of features}

\item{...}{additional arguments for \code{\link{wfs_request}}}
}
\value{
diversity summaries
}
\description{
Retrieves diversity summaries for a hexagonal grid
}

