% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occurrences.R
\name{get_occurrences}
\alias{get_occurrences}
\title{Occurrences}
\usage{
get_occurrences(identifiers, idColumn = "valid_aphia_id", filter = NULL,
  where = NULL, ...)
}
\arguments{
\item{identifiers}{vector of identifiers, these can be species names, AphiaIDs, OBIS ids, or storedpaths}

\item{idColumn}{which column to use when the identifier is an id, one of \code{valid_aphia_id}, \code{valid_id}}

\item{filter}{list of filters}

\item{where}{explicit where clause}

\item{...}{additional arguments for \code{\link{wfs_request}}}
}
\value{
occurrence data
}
\description{
Retrieves occurrence data
}

