% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_distribution.R
\name{map_distribution}
\alias{map_distribution}
\title{Distribution map}
\usage{
map_distribution(data, style = style_distribution())
}
\arguments{
\item{data}{data frame from \code{\link{get_distribution}}}

\item{style}{list with breaks, labels, and colors vectors}
}
\description{
Creates a map from distribution data
}

