% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Create a leaflet map.}
\usage{
map_leaflet(data, color = "#ff3399",
  provider_tiles = "OpenStreetMap.BlackAndWhite", popup = function(x) { x["id"] }, antarctic = FALSE)
}
\arguments{
\item{data}{the occurrences from \code{occurrence()}.}

\item{color}{color to be used for the dots.}

\item{provider_tiles}{the base map provider.}

\item{popup}{function generating the popup content.}

\item{antarctic}{use antarctic polar stereographic projection.}
}
\description{
Create a leaflet map.
}
