% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{Fetch a list of nodes}
\usage{
node(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Optional parameter to enable verbose logging (default = \code{FALSE}).}
}
\value{
The nodes
}
\description{
Fetch a list of nodes
}
\examples{
nodes <- node()
}
