% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obisclient.R
\name{occurrence}
\alias{occurrence}
\title{Find occurrences.}
\usage{
occurrence(scientificname = NULL, year = NULL, obisid = NULL,
  aphiaid = NULL, startdate = NULL, enddate = NULL, geometry = NULL,
  verbose = FALSE)
}
\arguments{
\item{scientificname}{}

\item{year}{}

\item{obisid}{}

\item{aphiaid}{}

\item{startdate}{}

\item{enddate}{}

\item{geometry}{A wkt geometry string.}
}
\value{
The occurrences.
}
\description{
Find occurrences.
}
\examples{
occurrence(scientificname = "Abra sibogai")
}

