% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\name{occurrence}
\alias{occurrence}
\title{Find occurrences.}
\usage{
occurrence(scientificname = NULL, taxonid = NULL, datasetid = NULL,
  nodeid = NULL, startdate = NULL, enddate = NULL,
  startdepth = NULL, enddepth = NULL, geometry = NULL,
  verbose = FALSE)
}
\description{
Find occurrences.
}
