% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occurrence.R
\name{occurrence}
\alias{occurrence}
\title{Find occurrences.}
\usage{
occurrence(scientificname = NULL, year = NULL, obisid = NULL,
  aphiaid = NULL, startdate = NULL, enddate = NULL, geometry = NULL,
  qc = NULL, verbose = FALSE)
}
\arguments{
\item{scientificname}{}

\item{year}{}

\item{obisid}{}

\item{aphiaid}{}

\item{startdate}{}

\item{enddate}{}

\item{geometry}{A wkt geometry string.}

\item{qc}{A vector of quality control flags you want to filter on. List of \link[=qc]{QC flags}.}
}
\value{
The occurrence records.
}
\description{
Find occurrences.
}
\examples{
occurrence(scientificname = "Abra sibogai")
occurrence(scientificname = "Abra sibogai", qc = c(1:6, 27))
}

