% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robis-package.R
\docType{package}
\name{robis}
\alias{robis-package}
\alias{robis}
\title{R client for the OBIS API}
\description{
This is an R client for the Ocean Biodiversity Information System (OBIS) API,
documented at \url{https://api.obis.org}.
}
\details{
Code and issues are available at \url{https://github.com/iobis/robis}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iobis/robis}
  \item Report bugs at \url{https://github.com/iobis/robis/issues}
}

}
\author{
\strong{Maintainer}: Pieter Provoost \email{pieterprovoost@gmail.com} (\href{https://orcid.org/0000-0002-4236-0384}{ORCID})

Authors:
\itemize{
  \item Samuel Bosch \email{mail@samuelbosch.com} (\href{https://orcid.org/0000-0002-2514-0283}{ORCID})
}

Other contributors:
\itemize{
  \item Ward Appeltans (\href{https://orcid.org/0000-0002-3237-4547}{ORCID}) [contributor]
  \item Stephen Formel (\href{https://orcid.org/0000-0001-7418-1244}{ORCID}) [contributor]
  \item  OBIS \email{helpdesk@obis.org} [copyright holder]
}

}
\keyword{internal}
