% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon}
\alias{taxon}
\title{Fetch taxon.}
\usage{
taxon(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Optional parameter to enable verbose logging (default
= \code{FALSE}).}
}
\value{
The areas.
}
\description{
Fetch taxon.
}
\seealso{
\code{\link{occurrence}} \code{\link{checklist}} \code{\link{node}}
  \code{\link{area}} \code{\link{dataset}} \code{\link{group}}
}
