% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon}
\alias{taxon}
\title{Generate taxon lists.}
\usage{
taxon(scientificname = NULL, year = NULL, obisid = NULL, aphiaid = NULL,
  startdate = NULL, enddate = NULL, geometry = NULL, qc = NULL,
  verbose = FALSE)
}
\arguments{
\item{scientificname}{}

\item{year}{}

\item{obisid}{}

\item{aphiaid}{}

\item{startdate}{}

\item{enddate}{}

\item{geometry}{A wkt geometry string.}

\item{qc}{A vector of qc numbers you want to filter out.}
}
\value{
The taxon list.
}
\description{
Generate taxon lists.
}

